Adding pessoa and organizador class new teste

+++ src/main/java/br/com/mack/persistence/entities/Participante.java
@@ -10,7 +10,7 @@ public class Participante extends Pessoa implements Serializable{
     public Participante() {
     }
 
-    public Participante(String formacao, long id_pessoa,String nome, String email, String senha, int celular) {
+    public Participante(String formacao, long id_pessoa,String nome, String email, String senha, double celular) {
         super(id_pessoa,nome, email, senha, celular);
         this.formacao = formacao;
     }
